new component for container as need to watch open zones; resizing containers is looking good
