+++ src/App.vue
@@ -3,7 +3,7 @@
     <div id="menu">
       <side-menu />
     </div>
-    <div id="container">
+    <div id="main-grid">
       <!-- <widget-controller /> -->
       <container />
     </div>
@@ -14,18 +14,18 @@
 import Vue from "vue"
 import Component from "vue-class-component"
 import SideMenu from "@/views/SideMenu.vue"
-import Container from "@/views/Container.vue"
+import Container from "@/views/GridController.vue"
 
 @Component({
   components: {
     Container,
-    SideMenu
-  }
+    SideMenu,
+  },
 })
 export default class App extends Vue {}
 </script>
 
-<style lang="scss">
+<style lang="css">
 body {
   @apply transition-colors bg-gray-50 dark:bg-gray-800;
 }
@@ -33,12 +33,13 @@ body {
 #app {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
   display: grid;
-  grid-template-areas: "menu container";
-  grid-template-columns: 3rem 1fr;
-  overflow: hidden;
+  grid-template-areas: "menu main-grid";
+
+  /* adding padding instead of expanding/empty cols leads to weird effects when resizing containers */
+  grid-template-columns: 3.25rem 1fr 0.25rem;
+
+  @apply text-center overflow-hidden;
 }
 
 #menu {
@@ -48,8 +49,8 @@ body {
   @apply h-screen;
 }
 
-#container {
-  grid-area: container;
+#main-grid {
+  grid-area: main-grid;
   overflow: auto;
 
   @apply h-screen;
